Fix sending files over the size of int max value
When I changed int to long when storing the file size I forgot to change the value that has the size of the file that has already been written

+++ src/main/java/dev/juho/ffmpegrender/server/client/reader/FileData.java
@@ -32,7 +32,7 @@ public class FileData implements ReaderData {
 	private boolean isFull;
 
 	private BufferedOutputStream fileOS;
-	private int fileWritten;
+	private long fileWritten;
 
 
 	public FileData(String rootFolder) {
